pyproject.toml
[tool.poetry]
name = "beeai junior test"
version = "0.1.0"
description = ""
authors = ["chugunova24 <t.chugunova24@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
django = "^4.2.6"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
